split card styles into their own files
